requirements.txt
Flask==2.2.3
gTTS==2.2.3
pytesseract==0.3.10
pdfminer.six==20201018
python-docx==0.8.11
Pillow==8.3.2
